{"schema":"decodex/commit/1","summary":"Split MCP docs context resources","authority":"XY-1115"}
