+++ apps/decodex/src/mcp/resources/context/docs/listing.rs
@@ -1,0 +1,93 @@
+use std::{
+	fs,
+	io::ErrorKind,
+	path::{Path, PathBuf},
+};
+
+use crate::{
+	mcp::{McpError, resources::types::McpResource},
+	prelude::Result,
+};
+
+pub(super) fn docs_resources(repo_root: &Path) -> Result<Vec<McpResource>, McpError> {
+	let mut resources = Vec::new();
+
+	push_file_resource(
+		&mut resources,
+		repo_root.join("docs/index.md"),
+		"decodex://docs/index",
+		"Documentation index",
+		"Checked-in Decodex documentation router.",
+	);
+	push_file_resource(
+		&mut resources,
+		repo_root.join("docs/policy.md"),
+		"decodex://docs/policy",
+		"Documentation policy",
+		"Checked-in Decodex documentation policy.",
+	);
+
+	for lane in ["spec", "runbook", "reference", "decisions", "research"] {
+		let docs_dir = repo_root.join("docs").join(lane);
+
+		for entry in read_sorted_dir(&docs_dir)? {
+			let Some(stem) = markdown_stem(&entry) else {
+				continue;
+			};
+
+			resources.push(McpResource::markdown(
+				format!("decodex://docs/{lane}/{stem}"),
+				format!("docs/{lane}/{stem}.md"),
+				"Checked-in Decodex documentation resource.",
+			));
+		}
+	}
+	for entry in read_sorted_dir(&repo_root.join("docs/research"))? {
+		let Some(stem) = markdown_stem(&entry) else {
+			continue;
+		};
+
+		resources.push(McpResource::markdown(
+			format!("decodex://research/{stem}"),
+			format!("docs/research/{stem}.md"),
+			"Checked-in Markdown Research Contract concept.",
+		));
+	}
+
+	Ok(resources)
+}
+
+fn push_file_resource(
+	resources: &mut Vec<McpResource>,
+	path: PathBuf,
+	uri: &str,
+	name: &str,
+	description: &str,
+) {
+	if path.is_file() {
+		resources.push(McpResource::markdown(uri, name, description));
+	}
+}
+
+fn read_sorted_dir(path: &Path) -> Result<Vec<PathBuf>, McpError> {
+	let entries = match fs::read_dir(path) {
+		Ok(entries) => entries,
+		Err(error) if error.kind() == ErrorKind::NotFound => return Ok(Vec::new()),
+		Err(error) => return Err(McpError::internal(error)),
+	};
+	let mut paths = entries
+		.map(|entry| entry.map(|entry| entry.path()).map_err(McpError::internal))
+		.collect::<Result<Vec<_>, _>>()?;
+
+	paths.sort();
+
+	Ok(paths)
+}
+
+fn markdown_stem(path: &Path) -> Option<String> {
+	if path.extension().and_then(|extension| extension.to_str()) != Some("md") {
+		return None;
+	}
+
+	path.file_stem().and_then(|stem| stem.to_str()).map(str::to_owned)
+}
